Move execute functions into their own modules
